### 2019-08-30 #1694 @whedon
[PRE REVIEW]: DEPP - Differential Evolution Parallel Program
**Submitting author:** @gbertoldo (<a href="http://orcid.org/0000-0002-1888-9181">Guilherme Bertoldo</a>)
**Repository:** <a href="https://github.com/gbertoldo/DEPP" target ="_blank">https://github.com/gbertoldo/DEPP</a>
**Version:** v1.0.0
