### 2019-08-30 #1694 @labarba
@gbertoldo — I think you need to remove the backticks around the citations, so the references will get processed when compiling the paper. After you push the changes, you can run `@whedon generate pdf` alone in a comment here to compile the paper again.
