### 2019-08-30 #1694 @labarba
👋  @meg-simula — Maybe this could be a good paper for you to edit? Have a look and let me know if you can lead this one.
"DEPP - Differential Evolution Parallel Program" https://github.com/gbertoldo/DEPP
