+++ paper.bib
@@ -25,19 +25,19 @@
 
 @Article{Vincenzi:2015,
   author               = {Vincenzi, Loris and Savoia, Marco},
-  title                = {{Coupling Response Surface and Differential Evolution for Parameter Identification Problems}},
-  journal              = {{COMPUTER-AIDED CIVIL AND INFRASTRUCTURE ENGINEERING}},
-  year                 = {{2015}},
-  volume               = {{30}},
-  number               = {{5, SI}},
-  pages                = {{376-393}},
-  month                = {{MAY}},
-  issn                 = {{1093-9687}},
-  doi                  = {{10.1111/mice.12124}},
-  eissn                = {{1467-8667}},
-  orcid-numbers        = {{Vincenzi, Loris/0000-0003-2541-7104}},
-  researcherid-numbers = {{Vincenzi, Loris/O-1975-2015}},
-  unique-id            = {{ISI:000352790800005}}
+  title                = {Coupling Response Surface and Differential Evolution for Parameter Identification Problems},
+  journal              = {COMPUTER-AIDED CIVIL AND INFRASTRUCTURE ENGINEERING},
+  year                 = {2015},
+  volume               = {30},
+  number               = {5, SI},
+  pages                = {376-393},
+  month                = {MAY},
+  issn                 = {1093-9687},
+  doi                  = {10.1111/mice.12124},
+  eissn                = {1467-8667},
+  orcid-numbers        = {Vincenzi, Loris/0000-0003-2541-7104},
+  researcherid-numbers = {Vincenzi, Loris/O-1975-2015},
+  unique-id            = {ISI:000352790800005}
 }
 
 @Book{Freeman:2004,
