+++ paper.bib
@@ -53,10 +53,3 @@
   year   = {2019},
   url    = {https://github.com/gbertoldo/DEPP},
 }
-
-@Misc{Radtke2019b,
-  author = {Radtke, J J and Bertoldo, G and Marchi, C H},
-  title  = {DEPP documentation},
-  year   = {2019},
-  url    = {https://github.com/gbertoldo/DEPP/files/3335317/depp.pdf},
-}
