### 2019-08-31 #1694 @HaoZeke
Hi @meg-simula. May I put myself forward for this? I really like modern fortran :grin:
