### 2019-09-03 #1694 @meg-simula
I've contacted David Ham (via email) to also review this paper, and he has agreed. I'll add him as an reviewer and let whedon start the review now.
