### 2019-09-16 gbertoldo/DEPP#2 @HaoZeke
[JOSS] Write up
Hi, as part of https://github.com/openjournals/joss-reviews/issues/1701, I would prefer if some more of the algorithms could be presented in the JOSS paper. I notice that the `depp_theory.pdf` file has a much better explanation of the overall flow. Could the `.md` file be made to be closer to it?
