### 2019-09-16 gbertoldo/DEPP#3 @HaoZeke
[JOSS] Test-suite
The examples are nice, but the `README.md` needs to be clearer about running tests. Actually, though not crucial, it would be nice to have a set of automated tests.

https://github.com/openjournals/joss-reviews/issues/1701
