### 2019-09-16 gbertoldo/DEPP#4 @HaoZeke
[JOSS] GPL Headers
Part of https://github.com/openjournals/joss-reviews/issues/1701. As per the GPL, please add  headers to the code.
![image](https://user-images.githubusercontent.com/4336207/64956755-f4856900-d8a8-11e9-99ff-f919aacea2fe.png)

I can open a pull request with the headers added in a few days.
