### 2019-10-01 #1701 @HaoZeke
I have opened some issues which have not yet been resolved. I could
complete the rest of the review but it will not be complete until the
authors address the issues.
