### 2019-10-02 #1701 @gbertoldo
Dear @meg-simula  and @HaoZeke, I was not sure about modifying the repository during the review. I will provide changes asked by @HaoZeke  as soon as possible.
