### 2019-10-02 0571751
GPL headers added to source code.
