### 2019-10-02 9eef4a2
Added a script for automated test.
