### 2019-10-02 gbertoldo/DEPP#3 @gbertoldo
I added a Bash script autoCheck.sh in the root of repository. This script checks the compilation of the source code and checks the execution of an example. I also updated the wiki page with instructions on how to used autoCheck.sh (see https://github.com/gbertoldo/DEPP/wiki#how-to-install-depp)
