### 2019-10-02 #1701 @HaoZeke
@gbertoldo, sorry I should have clarified the process at the onset. JOSS is an inclusive journal, and as the review progresses, I shall be opening issues to shepard the software towards being in accordance to our guidelines. Once the issues have all been addressed, @meg-simula will give her comments and move forward with the publication. Thanks for getting started on the issues.
