### 2019-10-02 gbertoldo/DEPP#3 @HaoZeke
It would be best to give a list of the dependencies required along with the version which shipped with the linux distro at the time.
