### 2019-10-03 b156b63
First commit of a more complete JOSS paper.
