### 2019-10-03 6409b3b
Removed first version of JOSS paper.
