+++ paper.bib
@@ -1,55 +1,0 @@
-@Book{Griva:2009,
-  Title                    = {Linear and Nonlinear Optimization},
-  Author                   = {Griva, I and Nash, S G and Sofer, A},
-  Publisher                = {SIAM},
-  Year                     = {2009},
-  Address                  = {USA},
-  Edition                  = {2}
-}
-
-@Book{Feoktistov:2006,
-  Title                    = {Differential evolution: in search of solutions},
-  Author                   = {Feoktistov, Vitaliy},
-  Publisher                = {Springer Science+Business Media},
-  Year                     = {c2006},
-  Address                  = {New York}
-}
-
-@Book{Coley:1999,
-  Title                    = {An Introduction to Genetic Algorithms for Scientists and Engineers},
-  Author                   = {Coley, D A},
-  Publisher                = {World Scientific},
-  Year                     = {1999},
-  Address                  = {Singapore}
-}
-
-@Article{Vincenzi:2015,
-  author               = {Vincenzi, Loris and Savoia, Marco},
-  title                = {Coupling Response Surface and Differential Evolution for Parameter Identification Problems},
-  journal              = {COMPUTER-AIDED CIVIL AND INFRASTRUCTURE ENGINEERING},
-  year                 = {2015},
-  volume               = {30},
-  number               = {5, SI},
-  pages                = {376-393},
-  month                = {MAY},
-  issn                 = {1093-9687},
-  doi                  = {10.1111/mice.12124},
-  eissn                = {1467-8667},
-  orcid-numbers        = {Vincenzi, Loris/0000-0003-2541-7104},
-  researcherid-numbers = {Vincenzi, Loris/O-1975-2015},
-  unique-id            = {ISI:000352790800005}
-}
-
-@Book{Freeman:2004,
-  Title                    = {Head First Design Patterns},
-  Author                   = {Freeman, Eric and Freeman, Elisabeth and Bates, Bert and Sierra, Kathy},
-  Publisher                = {O'Reilly Media},
-  Year                     = {2004}
-}
-
-@Misc{Radtke:2019a,
-  author = {Radtke, J J and Bertoldo, G and Marchi, C H},
-  title  = {DEPP: Differential Evolution Parallel Program},
-  year   = {2019},
-  url    = {https://github.com/gbertoldo/DEPP},
-}
