### 2019-10-03 f566db0
Editing paper...
