+++ JOSSPaper/paper.bib
@@ -51,5 +51,137 @@
   author = {Radtke, J J and Bertoldo, G and Marchi, C H},
   title  = {DEPP: Differential Evolution Parallel Program},
   year   = {2019},
-  url    = {https://github.com/gbertoldo/DEPP},
+  url    = {https://github.com/gbertoldo/DEPP}
+}
+
+@Article{Storn:1997,
+  Title                    = {Differential Evolution -- A Simple and Efficient Heuristic for global Optimization over Continuous Spaces},
+  Author                   = {Storn, Rainer
+and Price, Kenneth},
+  Journal                  = {Journal of Global Optimization},
+  Year                     = {1997},
+  Month                    = {Dec},
+  Number                   = {4},
+  Pages                    = {341--359},
+  Volume                   = {11},
+  Doi                      = {10.1023/A:1008202821328},
+  ISSN                     = {1573-2916}
+}
+
+@Book{Price:2005,
+  Title                    = {Differential evolution: a practical approach to global optimization},
+  Author                   = {Price, K V and Storn, R M and Lampinen, J A},
+  Publisher                = {Springer},
+  Year                     = {2005},
+  Address                  = {Germany}
+}
+
+@Book{Khuri:1996,
+  Title                    = {{R}esponse {S}urface: {D}esign and {A}nalyses},
+  Author                   = {Khuri, A. I. and Cornell, J. A.},
+  Publisher                = {Marcel Dekker,Inc},
+  Year                     = {1996},
+  Address                  = {New York},
+  Edition                  = {2},
+  Series                   = {Statistics, textbooks and monographs},
+  Volume                   = {152}
+}
+
+@Book{Myers:2009,
+  Title                    = {Response surface methodology: process and product optimization using designed experiments},
+  Author                   = {Myers, Raymond H and Montgomery, Douglas C and Anderson-Cook, Christine M},
+  Publisher                = {Wiley},
+  Year                     = {c2009},
+  Address                  = {Hoboken, N.J.},
+  Edition                  = {3}
+}
+
+@Book{Brainerd:2015,
+  Title                    = {Guide to {F}ortran 2008 {P}rogramming},
+  Author                   = {Brainerd, Walter S.},
+  Publisher                = {Springer-Verlag},
+  Year                     = {2015},
+  Address                  = {London},
+  Edition                  = {2}
+}
+
+@TechReport{Tennessee:2009,
+  Title                    = {{MPI}: {A} {M}essage-{P}assing {I}nterface {S}tandard},
+  Institution              = {University of Tennessee},
+  Year                     = {2009}
+}
+
+@Manual{Ramey:2016,
+  Title                    = {Bash Reference Manual},
+  Author                   = {Ramey, Chet and Fox, Brian},
+  Edition                  = {4.4},
+  Month                    = {September},
+  Organization             = {Case Western Reserve University and Free Software Foundation},
+  Year                     = {2016}
+}
+
+@Book{Freeman:2004,
+  Title                    = {Head First Design Patterns},
+  Author                   = {Freeman, Eric and Freeman, Elisabeth and Bates, Bert and Sierra, Kathy},
+  Publisher                = {O'Reilly Media},
+  Year                     = {2004}
+}
+
+
+
+
+
+
+
+
+
+
+
+
+
+@Book{Haupt:2004,
+  Title                    = {Practical genetic algorithms},
+  Author                   = {Haupt, R L and Haupt, S E},
+  Publisher                = {John Wiley},
+  Year                     = {c2004},
[…]
