### 2019-10-03 4f9de53
Editing paper...
