+++ JOSSPaper/paper.md
@@ -169,4 +169,112 @@ Finally, the mutant vector is obtained by maximization of the response surface.
 
 If the application of RSM fails or the trial individual is out of range, then a pure DE individual is created.
 
+Examples {#sec:examples}
+========
+
+Four test functions, whose maximizers are known analytically, were chosen for illustrating the application of DEPP:
+
+-   Step function [@Zhang:2009] $f(\boldsymbol{x})=-\sum_{i=1}^{D} \lt( \lfloor x_i-0.5 \rfloor \rt)^2$,
+![Step function.](function02.png){width="50.00000%"}
+
+-   Rosenbrock function [@Feoktistov:2006] $f(\boldsymbol{x})=-\sum_{i=1}^{D-1} 100\lt( x_i^2-x_{i+1}\rt)^2+(1-x_i)^2$,
+![Rosenbrock function.](function09.png){width="50.00000%"}
+
+-   Noisy Quartic function [@Zhang:2009] $f(\boldsymbol{x})=-\mathcal{R}-\sum_{i=1}^{D} i~x_i^4$, where $\mathcal{R}$ is a uniform
+    random number belonging to $[0,1)$,
+![Noisy Quartic function.](function13.png){width="50.00000%"}
+
+-   Schwefel 2.26 function [@Zhang:2009] $f(\boldsymbol{x})=-418.98288727243369~D+\sum_{i=1}^{D} x_i\sin{\lt(\sqrt{|x_i|}\rt)}$.
+![Schwefel 2.26 function.](function14.png){width="50.00000%"}
+
+These functions were chosen because they present particular features that make optimization difficult. Step function (Fig. \[fig:step\]), a set o flat surfaces, pose a hard problem to optimizers relying on functions’ derivatives. This is even worst for Noisy Quartic function (Fig. \[fig:noisy\]), which has a noisy plateau near its maximum. Rosenbrock function (Fig. \[fig:rosenbrock\]), although smooth, has a sharp narrow ridge around a parabola, that makes it a challenge to find
+an appropriate search direction. Finally, Schwefel 2.26 (Fig. \[fig:schwefel\]) has several local optima, which may lead to premature convergence.
+
+The maximizers $x_i^*$ and the domain of optimization of the test functions are presented in
+Table \[tab:boundsmaximizer\].
+
+     Function      $L_i$   $U_i$       $x_i^*$
+  --------------- ------- ------- ------------------
+       Step        -100     100          0.5
+    Rosenbrock      -2       2            1
+   Noisy Quartic   -1.28   1.28           0
+   Schwefel 2.26   -500     500    420.968597844358
+
+  : Lower $L_i$ and upper $U_i$ bounds for maximization and the global maximizer $x_i^*$ ($1\le i \le D$).
+
+\[tab:boundsmaximizer\]
+
+Additionally, this example aims (i) to show the effect of DE-Response Surface (DE-RSM) coupling on reducing the number of iterations required to achieve convergence and (ii) show the effect of parallelization on reducing the optimization time. In order to accomplish the first task, test functions (\[eq:step\])-(\[eq:schwefel\]) were optimized for $D \in \{2, 4, 8\}$ dimensions using DE and DE-RSM. For the second task, the number of dimensions was restricted to $D=2$, but for each call of the external software that calculates the objective function it was added a delay of 1 s in order to simulate a computationally expensive program. Since DE is a stochastic method, in both tasks, the final results are the average of 50 optmizations.
+
[…]
