+++ JOSSPaper/paper.bib
@@ -26,7 +26,7 @@
 @Article{Vincenzi:2015,
   author               = {Vincenzi, Loris and Savoia, Marco},
   title                = {Coupling Response Surface and Differential Evolution for Parameter Identification Problems},
-  journal              = {COMPUTER-AIDED CIVIL AND INFRASTRUCTURE ENGINEERING},
+  journal              = {Computer-Aided Civil And Infrastructure Engineering},
   year                 = {2015},
   volume               = {30},
   number               = {5, SI},
@@ -127,56 +127,10 @@ and Price, Kenneth},
   Year                     = {2004}
 }
 
-
-
-
-
-
-
-
-
-
-
-
-
-@Book{Haupt:2004,
-  Title                    = {Practical genetic algorithms},
-  Author                   = {Haupt, R L and Haupt, S E},
-  Publisher                = {John Wiley},
-  Year                     = {c2004},
-  Address                  = {Hoboken, N.J.},
-  Edition                  = {2}
-}
-
-
-
-@InProceedings{Xiao:2014,
-  author            = {Xiao, Chixin and Xue, Zhigang and Xiao, Chixin and Yin, Jianping},
-  title             = {{Rational Models to Improve Performance of Differential Evolution for MOEA/D}},
-  booktitle         = {{2014 10TH INTERNATIONAL CONFERENCE ON NATURAL COMPUTATION (ICNC)}},
-  year              = {{2014}},
-  series            = {{Proceedings International Conference on Natural Computation}},
-  pages             = {{335-342}},
-  note              = {{10th International Conference on Natural Computation (ICNC), Xiamen, PEOPLES R CHINA, AUG 19-21, 2014}},
-  book-group-author = {{IEEE}},
-  isbn              = {{978-1-4799-5151-2}},
-  issn              = {{2469-8814}},
-  orcid-numbers     = {{Xiao, Chixin/0000-0002-3911-7113}}
-}
-
-
-
-
-
-
-
-
-
-
 @Article{Zhang:2009,
   author        = {Zhang, Jingqiao and Sanderson, Arthur C.},
   title         = {{JADE: Adaptive Differential Evolution With Optional External Archive}},
-  journal       = {{IEEE TRANSACTIONS ON EVOLUTIONARY COMPUTATION}},
+  journal       = {{IEEE Transactions on Evolutionary Computation}},
   year          = {{2009}},
   volume        = {{13}},
   number        = {{5}},
@@ -185,3 +139,4 @@ and Price, Kenneth},
   issn          = {{1089-778X}},
   doi           = {{10.1109/TEVC.2009.2014613}}
 }
+
