### 2019-10-03 8db898b
Editing paper...
