### 2019-10-05 3e913f1
Editing paper...
