### 2019-10-05 b53ccc4
Editing paper...
