### 2019-10-05 55340f3
Editing paper...
