### 2019-10-05 26f6f96
Editing paper...
