### 2019-10-05 c3611b9
Editing paper...
