### 2019-10-06 2e599d9
Editing paper...
