+++ JOSSPaper/paper.md
@@ -1,4 +1,11 @@
 ---
+title: "DEPP"
+author: "Guilherme"
+date: "05/10/2019"
+output: pdf_document
+---
+
+---
 title: 'DEPP - Differential Evolution Parallel Program'
 tags:
   - Optimization
@@ -48,23 +55,22 @@ Software Architecture {#sec:arch}
 DEPP is written in FORTRAN 2008 standard language[@Brainerd:2015] with the Message Passing Interface (MPI) directives[@Tennessee:2009] and takes
 advantage of the Object-Oriented Paradigm.
 
-The folder structure of DEPP is shown in Figure \[fig:depptree\]. The directory *depp\_input* contains the input (text) files which defines
+The folder structure of DEPP is shown in Figure 1. The directory *depp\_input* contains the input (text) files which defines
 the control parameters of the optimization. Results of DEPP are saved into *depp\_output* directory. The source code is within *depp\_src* directory. This directory also contains the Bash[@Ramey:2016] script *compile.sh*, which compiles DEPP source code and generates the executable *depp.exe* in the root of the file structure.  The interface between DEPP and the external program (EP), which calculates the objective function, is defined in the *interface* directory. Finally, the script *run.sh* runs DEPP using MPI.
 
 ![Folder structure of DEPP.](folder_structure.png){width="50.00000%"}
-[\[fig:depptree\]]{}
+
 
 Thanks to the Object-Oriented Programming, the source code was designed to simplify the implementation of new methods. The algorithm works on
-abstract classes separated, basically, in (i) population initializers, (ii) search strategies, (iii) fitness calculation and (iv) stop conditions. The concrete classes are generated using the Factory Design Pattern[@Freeman:2004]. All MPI commands are encapsulated into proper classes, following the logic of the Adapter Design Pattern. In this way, users interested in the implementation of new methods will not concern about MPI details.
+abstract classes separated, basically, as (i) population initializers, (ii) search strategies, (iii) fitness calculation and (iv) stop conditions. The concrete classes are generated using the Factory Design Pattern[@Freeman:2004]. All MPI commands are encapsulated into proper classes, following the logic of the Adapter Design Pattern. In this way, users interested in the implementation of new methods will not concern about MPI details.
 
 Interfacing DEPP with an external program is exemplified in *interface* directory. This folder contains three examples based on C++ and FORTRAN 2008 languages, although many other programming languages may be used for interfacing.
 
-The basic algorithm of DEPP is illustrated in Figure \[fig:deppalgorithm\]. DEPP reads the input data from *depp\_input* directory (optionally, it may start from some backup). The stop condition is analyzed. In the first iteration, the stop condition is generally not satisfied (it may not be true when starting from a
+The basic algorithm of DEPP is illustrated in Figure 2. DEPP reads the input data from *depp\_input* directory (optionally, it may start from some backup). The stop condition is analyzed. In the first iteration, the stop condition is generally not satisfied (it may not be true when starting from a
[…]
