### 2019-10-06 0cffd01
Editing paper...
