### 2019-10-06 gbertoldo/DEPP#2 @gbertoldo
Hi, @HaoZeke. We reproduced depp_theory.pdf in JOSS paper.
