+++ interface/Functions/compile.sh
@@ -1,3 +1,40 @@
+#    DEPP - Differential Evolution Parallel Program
+#
+#    Copyright (C) 2019 by Jonas Joacir Radtke, Guilherme Bertoldo and Carlos Henrique Marchi
+#
+#    This program is free software: you can redistribute it and/or modify
+#    it under the terms of the GNU General Public License as published by
+#    the Free Software Foundation, either version 3 of the License, or
+#    (at your option) any later version.
+#
+#    This program is distributed in the hope that it will be useful,
+#    but WITHOUT ANY WARRANTY; without even the implied warranty of
+#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+#    GNU General Public License for more details.
+#
+#    You should have received a copy of the GNU General Public License
+#    along with this program.  If not, see <http://www.gnu.org/licenses/>.
+#    
+#    Contact:
+#          Jonas Joacir Radtke (a)
+#                 E-mail: jonas.radtke@gmail.com
+#
+#          Guilherme Bertoldo (a)
+#                 E-mail: glbertoldo@gmail.com
+#
+#          Carlos Henrique Marchi (b)
+#                 E-mail: chmcfd@gmail.com
+#    Institution
+#          (a) Federal University of Technology - Paraná - UTFPR
+#              Linha Santa Bárbara, s/n, Francisco Beltrão, Paraná, Brazil
+#              Zip Code 85601-970
+#              
+#          (b) Federal University of Paraná - UFPR
+#              Curitiba, Paraná, Brazil
+#              Caixa postal 19040
+#              Zip Code 81531-980
+#
+
 gfortran -std=f2008 -Wall -O3 -o fitness.x depp_interface.f90 functions.f90
 
 rm *.mod
