### 2019-10-06 gbertoldo/DEPP#3 @gbertoldo
Hi, @HaoZeke, I added headers to bash scripts. I also put the version of compilers into autoCheck.sh. We will review wiki soon.
