### 2019-10-07 #1701 @meg-simula
Hm. I assume you tried:

" If you cannot edit the checklist please:

   2. Be sure to accept the invite at this URL:
   https://github.com/openjournals/joss-reviews/invitations

"
?
