### 2019-10-08 2514ef8
Editing paper...
