### 2019-10-10 gbertoldo/DEPP#5 @dham
[JOSS] documentation
Part of https://github.com/openjournals/joss-reviews/issues/1701

The documentation seems a long way short of what would be needed by a new user. There doesn't really seem to be a manual, just two worked examples. There is also an automatically generated class documentation page, but this seems totally opaque. It's not at all clear what benefit a user would gain from attempting to read this.

The first worked example states "In general, inputs are self-explanatory." This is not true for me at least. How does a user know which options are available? Which values are legal for these options? What is their effect?

What is the required interface of an external function? What are the constraints on its semantics? For example what if the external function itself expects to run in parallel, how are communicators handled or shared?
