### 2019-10-10 gbertoldo/DEPP#6 @dham
[JOSS] no effective testing
Part of openjournals/joss-reviews#1701

The code does not have an adequate testing arrangement. What is provided is one example, and a script which appears to only check that that one example executes without crashing.

I would expect to see a suite of tests designed to show that the algorithm performs as expected for a variety of problems. As a user, I would expect to be able to establish from running the tests that the implementation is correct,  for example because a number of problems with known solutions are solved and the scripts test the correctness of the solutions.

Ideally I would expect to see continuous integration testing in order to ensure that changes to the software do not break it. At a minimum I would expect a command which can be run to execute all the tests, and report which of these pass and fail.
