### 2019-10-14 #1701 @meg-simula
@gbertoldo @HaoZeke and @dham Thanks for starting the review and starting addressing the issues.

@HaoZeke Has @gbertoldo addressed your initial concerns? If so, are you ready to proceed with the review?
