### 2019-10-14 gbertoldo/DEPP#5 @gbertoldo
Hello, @dham,

Thank you for your contributions.

There was a document in the repository describing the theory behind DEPP. I moved the content of this document to JOSS paper after a suggestion of @HaoZeke. I believe that the paper is fundamental to understand how to use DEPP. So, I will provide a link to the JOSS paper in the wiki.

The automatically generated class documentation page is intended for more advanced users, who want to modify the source code. I do not think it must be changed.

We will provide a more detailed explanation about input parameters and interfacing on the wiki page. We will also answer the question about the external function running in parallel.

Let me know if these modifications are satisfactory.
