### 2019-10-14 gbertoldo/DEPP#5 @dham
Thanks, I saw the change you made to the paper at the other reviewer's request. I think the key distinction that I am raising here is the lack of high quality user documentation, which is a different problem from the explanation of the mathematics underpinning the tool. In essence the issue is that there isn't really a manual in the sense that a user of software would expect. I think that my specific questions above are particular symptoms of this. The more generic question that a manual would need to answer is, how would a user with no connection to the authors know what all of the capabilities of the software are, and how to practically access those capabilities?

I also have some concerns about the software engineering of DEPP itself. I was waiting to confirm with the editor, @meg-simula, whether these were in scope for a JOSS review. Having learned that they are, I will open more issues on these subjects.
