### 2019-10-14 gbertoldo/DEPP#7 @dham
[JOSS] hard coded directories
Part of https://github.com/openjournals/joss-reviews/issues/1701

DEPP has hard-coded input and output directories. This is out of line with accepted conventions for software and makes it hard for DEPP to be used in scripts or as part of a reproducible science workflow.

The near-universal convention would be for the name of the input file to be a command-line argument to DEPP so that the input files can live wherever the user needs them to. The filename of external functions and the output files can be something which is specified in the input files, but this should be by default relative to the input file, so that the output files are in the same place relative to the input files, not relative to the location of the DEPP install.
