### 2019-10-14 gbertoldo/DEPP#5 @gbertoldo
Ok, I understand. We want to improve the software quality and your contributions are welcome. Could you give me examples of what is considered a high-quality user documentation?
