### 2019-10-14 gbertoldo/DEPP#5 @dham
I think a very good example is the documentation for triangle, which is a 2D mesh generator. https://www.cs.cmu.edu/~quake/triangle.html

Even though it's a different area of computational science, I think it's similar in scope to DEPP in that it's a single executable, compiled programme with something like the same number of optional parameters. The web page is a bit old-fashioned in style, but the content is very good. As a new user, I can easily find everything that triangle can do, and what options I would use to make triangle do it.
