### 2019-10-29 170444f
Now file of input parameters is passed by argument to DEPP.
