+++ examples/Fortran2008/compile.sh
@@ -35,6 +35,6 @@
 #              Zip Code 81531-980
 #
 
-gfortran -std=f2008 -Wall -O3 -o fitness.x depp_interface.f90 functions.f90
+gfortran -std=f2008 -Wall -O3 -o fitness.x Fortran2008.f90
 
 rm *.mod
