+++ examples/Fortran2008/depp_de_rsm_parameters.txt
@@ -1,27 +1,0 @@
-===========================================================================================================================================
-
-                                              DEPP DE-RSM CONFIGURATION FILE
-                                              
-All data must follow the input pattern: "value & variable & description", where '&' is the separation field delimiter.
-
--------------------------------------------------------------------------------------------------------------------------------------------
-
-                    DE-RSM &                      CID  &  Class ID of this configuration file
-    depp_de_parameters.txt &  de_search_strategy_conf  &  DE search strategy configuration file
-   depp_rsm_parameters.txt & rsm_search_strategy_conf  &  RSM search strategy configuration file
-                    0.35d0 &                       fh  &  Initial fraction of hybridization
-                    0.10d0 &                    fhmin  &  Minimum fraction of hybridization
-                    0.90d0 &                    fhmax  &  Maximum fraction of hybridization
-                         1 &                      fhm  &  Model for the dynamical calculation of the factor of hybridization
-
-===========================================================================================================================================
-
-Model for calculation of the factor of hybridization (fhm):
------------------------------------------------------------
-
-0 - fh = constant
-
-1 - fh = probability of success of the previous application of RSM
-
-===========================================================================================================================================
-

[…]
