### 2019-10-29 e8dd4e2
Code simplification.
