### 2019-10-30 #1701 @meg-simula
@gbertoldo Just checking in - how are matters going? Have you had a chance to address the issues raised by @dham @haozeke?
