### 2019-11-01 4bdca89
Removed hard coded directories and added a test suite.
