+++ examples/TestFunctions/input_function.txt
@@ -1,4 +1,4 @@
-          14 =        nf:   Function number
+          9 =        nf:   Function number
 
 ==============================================================================================
 NF    FUNCTION NAME                                    OBJECTIVE      DOMAIN
