### 2019-11-01 gbertoldo/DEPP#7 @gbertoldo
Hi, @dham

We modified DEPP to receive just one input file, that is passed as argument in the command-line. Indeed this change improved the usability of the software. We still need to update the wiki documentation and JOSS paper about this change.
