### 2019-11-01 01da3be
Renamed test suite folder.
