+++ html/namespacemod__class__optimizer.js
@@ -1,0 +1,4 @@
+var namespacemod__class__optimizer =
+[
+    [ "class_optimizer", "structmod__class__optimizer_1_1class__optimizer.html", "structmod__class__optimizer_1_1class__optimizer" ]
+];
