+++ doc/doxygen/html/search/classes_2.js
@@ -1,0 +1,4 @@
+var searchData=
+[
+  ['fit_5finterface',['fit_interface',['../interfacemod__class__abstract__rsm_1_1fit__interface.html',1,'mod_class_abstract_rsm']]]
+];
