+++ doc/doxygen/html/namespacemod__class__abstract__population__initializer.js
@@ -1,0 +1,5 @@
+var namespacemod__class__abstract__population__initializer =
+[
+    [ "class_abstract_population_initializer", "structmod__class__abstract__population__initializer_1_1class__abstract__population__initializer.html", "structmod__class__abstract__population__initializer_1_1class__abstract__population__initializer" ],
+    [ "get_trial_interface", "interfacemod__class__abstract__population__initializer_1_1get__trial__interface.html", null ]
+];
