### 2019-11-04 b594f1c
Organized test suite.
