+++ examples/TestFunctions/input_function.txt
@@ -1,4 +1,4 @@
-          9 =        nf:   Function number
+          1 =        nf:   Function number
 
 ==============================================================================================
 NF    FUNCTION NAME                                    OBJECTIVE      DOMAIN
