+++ examples/TestFunctions/depp_parameters.txt
@@ -6,11 +6,11 @@ All data must follow the input pattern: "value & variable & description", where 
 
 -------------------------------------------------------------------------------------------------------------------------------------------
 
-                                  function02 &                  sname  &  Simulation name
+                                  function01 &                  sname  &  Simulation name
                                            0 &              verbosity  &  Defines log verbosity level (0=min, 1=med, 2=max)
                                            0 &            save_backup  &  Saves backup (0=no, 1=yes: increases CPU time)
                                            0 &                 reload  &  Upload backup data and continue computation (0=no; 1=yes)
-                                        $PWD &                   fdir  &  Name of working directory for fitness calculation
+                                        $PWD &                   fdir  &  Name of the working directory for fitness calculation
                                    fitness.x &                   ffit  &  Name of executable for fitness calculation
                                            2 &                     nu  &  Number of unknowns
                                           20 &                     np  &  Size of population (recommended 10*nu)
@@ -24,7 +24,7 @@ Search strategy models:
 
 Population initializer models:
 ------------------------------
-                         - uniform_random: individuals of initial population are randomly generated                              
+                         - uniform_random: individuals of the initial population are randomly generated                              
 
 ===========================================================================================================================================
 
@@ -33,15 +33,15 @@ Population initializer models:
 -------------------------------------------------------------------------------------------------------------------------------------------
 
 Stop condition is a composite of the following models:
-                         - max_generations: stops if number of generations exceeds ng (set ng bellow)
-                         - no_improvement: stops if the best fitness is not improved within GNoAcc generations (set GNoAcc bellow)
-                         - p_measure: stops if p-measure of Feoktistov is satisfied within tolerance ptol (set ptol bellow)
+                         - max_generations: stops if the number of generations exceeds ng (set ng below)
+                         - no_improvement: stops if the best fitness is not improved within GNoAcc generations (set GNoAcc below)
+                         - p_measure: stops if p-measure of Feoktistov is satisfied within tolerance ptol (set ptol below)
 Choose a composition of models and set the corresponding parameters.
                          
   max_generations, no_improvement, p_measure & composite_stop_condition &  List of stop condition models
-                                       20000 &                      ng  &  Maximum number of generations (Model: max_generations)
-                                         500 &                  GNoAcc  &  Max. num. of gen. before stopping if no improvement was found (Model: no_improvement)
-                               dimensionless &                     kpm  &  Kind of population convergence measure (See options bellow) (Model: p_measure)
+                                         200 &                      ng  &  Maximum number of generations (Model: max_generations)
+                                          40 &                  GNoAcc  &  Max. num. of gen. before stopping if no improvement was found (Model: no_improvement)
+                               dimensionless &                     kpm  &  Kind of population convergence measure (See options below) (Model: p_measure)
                                     0.0005d0 &                    ptol  &  Tolerance for the population convergence measure (Model: p_measure)
 
 Kind of population convergence measure (kpm):
@@ -58,8 +58,8 @@ Kind of population convergence measure (kpm):
 Tip: if all components have the same value, just type one value, i.e., it is not necessary to repeat all components. Otherwise, type
 each component separated by comma, space or tab.
 
-                                     -500.d0 &             lower_bound  &  Components of the lower bound vector.
-                                      500.d0 &             upper_bound  &  Components of the upper bound vector.
+                                     -5.12d0 &             lower_bound  &  Components of the lower bound vector.
+                                      5.12d0 &             upper_bound  &  Components of the upper bound vector.
 
 
[…]
