### 2019-11-07 8015cac
Updated JOSS paper.
