### 2019-11-07 296f22e
Fixed typos and updated doxygen.
