+++ doc/doxygen/html/search/classes_3.js
@@ -1,6 +1,0 @@
-var searchData=
-[
-  ['get_5ffitness_5finterface',['get_fitness_interface',['../interfacemod__class__abstract__fitness__calculator_1_1get__fitness__interface.html',1,'mod_class_abstract_fitness_calculator']]],
-  ['get_5foptimizer_5finterface',['get_optimizer_interface',['../interfacemod__class__abstract__rsm_1_1get__optimizer__interface.html',1,'mod_class_abstract_rsm']]],
-  ['get_5ftrial_5finterface',['get_trial_interface',['../interfacemod__class__abstract__search__strategy_1_1get__trial__interface.html',1,'mod_class_abstract_search_strategy::get_trial_interface'],['../interfacemod__class__abstract__population__initializer_1_1get__trial__interface.html',1,'mod_class_abstract_population_initializer::get_trial_interface']]]
-];

[…]
