### 2019-11-07 5bbd28f
Fixed typos and updated doxygen.
