+++ html/search/functions_2.js
@@ -1,4 +1,0 @@
-var searchData=
-[
-  ['data_5fsize',['data_size',['../structmod__class__external__fitness__calculator_1_1class__external__fitness__calculator.html#ae8ffd7903595d147e953bfebd8838e2f',1,'mod_class_external_fitness_calculator::class_external_fitness_calculator::data_size()'],['../structmod__class__de__rsm_1_1class__de__rsm.html#ad1530b5d6a428f767558df56e6b975ed',1,'mod_class_de_rsm::class_de_rsm::data_size()'],['../structmod__class__rsm__search__strategy_1_1class__rsm__search__strategy.html#ab83b928ad0995274cb63082e60e16101',1,'mod_class_rsm_search_strategy::class_rsm_search_strategy::data_size()'],['../structmod__class__de__rand__1_1_1class__de__rand__1.html#ac80dde0885f1616e6f25f6eabc732c9b',1,'mod_class_de_rand_1::class_de_rand_1::data_size()']]]
-];

[…]
