### 2019-11-07 gbertoldo/DEPP#7 @gbertoldo
Hi, @dham and @HaoZeke, I think this issue is solved.
