### 2019-11-07 gbertoldo/DEPP#6 @gbertoldo
Hi, @dham and @HaoZeke, I think this issue is solved. There are two scripts within test/: one of them to check DEPP installation and another one for code verification. We also updated the Wiki documentation and the JOSS paper. We applied four test functions for code verification. These functions present different levels of difficulty to be optimized. They are also unimodal to ensure a unique answer when comparing numerical and analytical solutions. Since DEPP is a stochastic software, the solution will change from one run to another. However, numerical solutions must always agree with analytical one within the prescribed error tolerance. Instructions for code verification are given here: https://github.com/gbertoldo/DEPP/wiki/How-to-install-DEPP-on-Linux#code-verification.
