### 2019-11-07 gbertoldo/DEPP#5 @gbertoldo
Hi, @dham and @HaoZeke, I think this issue is solved. We updated the wiki page and JOSS paper.

We describe here https://github.com/gbertoldo/DEPP/wiki/How-to-install-DEPP-on-Linux how to install DEPP on six Linux distros, how to check DEPP installation and how to perform automatic code verification.
This page https://github.com/gbertoldo/DEPP/wiki/How-to-use-DEPP describes how to run DEPP using command-line terminal, how to perform DEPP interfacing with the external fitness calculator, explains how to set up the configuration file, presents the meaning of all parameters, their types and available options and, finally, how to use DEPP by examples.
