### 2019-12-07 #1701 @HaoZeke
@gbertoldo just to keep you informed. I will be working on my review with issues opened in the repository which should link back to this thread as well. Please feel free to address the issues as when and when you can.
