### 2019-12-07 gbertoldo/DEPP#8 @HaoZeke
[JOSS] DOIs Missing
Part of https://github.com/openjournals/joss-reviews/issues/1701
