### 2019-12-09 #1701 @gbertoldo
Thanks, @haoZeke!

Em sáb., 7 de dez. de 2019 às 02:45, Rohit Goswami <notifications@github.com>
escreveu:

> @gbertoldo <https://github.com/gbertoldo> just to keep you informed. I
> will be working on my review with issues opened in the repository which
> should link back to this thread as well. Please feel free to address the
> issues as when and when you can.
>
