### 2019-12-31 #1701 @HaoZeke
Hi, I just pulled changes, and I am wondering if there should be a version update? Or a tagged release, if the tests are complete?
