### 2020-01-09 #1701 @gbertoldo
Hi, @HaoZeke!

Current version is 1.0.0. Please, update the version number according to the following rules:

> "...form x.y.z, where "x" represents the major version, "y" represents the minor version, and "z" represents the number of the revision. For instance, version 1.3.22 would be major version 1, minor version 3, revision 22.
>
> - A major version number is incremented when there is a significant code change that might be incompatible with previous versions, such as a fundamental change of framework.
> - A minor version number is incremented when significant bug fixes are implemented, or a new feature is added.
> - A revision number is incremented when minor bug fixes are implemented."

This quote was extracted from https://www.computerhope.com/jargon/v/version.htm
