### 2020-02-07 #1701 @gbertoldo
Dear, @HaoZeke, @dham and @meg-simula,

I hope I finished the corrections. If there are other issues, please, let me know.
