### 2020-02-07 gbertoldo/DEPP#9 @HaoZeke
[JOSS] Contribution Guidelines
Part of https://github.com/openjournals/joss-reviews/issues/1701.
As per the checklist, please add a `CONTRIBUTING.md` or a similar file (or wiki section) describing the people to contact and method to collaborate (pull-requests, for example)
