### 2020-02-08 ad4dcd1
Added contributing guidelines.
