+++ CONTRIBUTING.md
@@ -1,0 +1,42 @@
+# Contributing
+
+Welcome to the contributing page! Here you will find some guidelines to help us improving DEPP.
+
+## Programming environment
+If you never used an Integrated Development Environment (IDE), we recommend you to install [CodeBlocks IDE](http://www.codeblocks.org/home). After install, you may try DEPP's CodeBlocks project available at /src/depp.cdb.
+
+## How to submit changes
+
+To submit changes, you must use a [fork](https://help.github.com/en/github/getting-started-with-github/fork-a-repo) and [pull request](https://help.github.com/en/github/collaborating-with-issues-and-pull-requests/creating-a-pull-request-from-a-fork) procedure.
+
+Before a pull request, you must:
+
+- Discuss the desired changes in the DEPP's [forum](https://groups.google.com/forum/#!forum/ask_depp);
+- Verify if the [installation script](https://github.com/gbertoldo/DEPP/wiki/How-to-install-DEPP-on-Linux#installation-check) is working properly;
+- Verify if the [test suite](https://github.com/gbertoldo/DEPP/wiki/How-to-install-DEPP-on-Linux#code-verification) is working properly;
+- Verify if the [examples](https://github.com/gbertoldo/DEPP/wiki/How-to-use-DEPP#examples) are working properly;
+- Update the version number;
+- Update the [wiki page](https://github.com/gbertoldo/DEPP/wiki), if necessary;
+- Update the README.md, if necessary.
+
+
+## Version number
+*(adapted from https://www.computerhope.com/jargon/v/version.htm)*
+
+DEPP's version number follows the pattern x.y.z, where "x" represents the major version, "y" represents the minor version, and "z" represents the number of the revision.
+
+- A major version number is incremented when there is a significant code change that might be incompatible with previous versions, such as a fundamental change of framework.
+- A minor version number is incremented when significant bug fixes are implemented, or a new feature is added.
+- A revision number is incremented when minor bug fixes are implemented.
+
+## How to report a bug and request an enhancement
+
+Bug reports and enhancement requests are done by raising [issues](https://github.com/gbertoldo/DEPP/issues).
+
+## Code of Conduct 
+
+All relationships must respect the code of conduct described at [Contributor Covenant page](https://www.contributor-covenant.org/version/2/0/code_of_conduct).
+
+
+
+
