### 2020-02-08 #1701 @gbertoldo
Hi, @HaoZeke! We added the CONTRIBUTING.md file at the root of the repository.
