### 2020-02-08 #1701 @HaoZeke
Excellent, @gbertoldo. @meg-simula I have completed my review and am satisfied that DEPP is suitable for publication in JOSS in it's current form. Apologies for the many delays!
