### 2020-03-02 #1701 @dham
OK, I'm now happy with this. In particular, the documentation and testing is vastly better than in the original submission.
