### 2020-03-10 #1701 @meg-simula
@gbertoldo: Nice work!

Could you please also address the following editorial comments for the paper:

- [ ] Please run a spell-check on the paper see e.g. "paralellization", "a set o flat"
- [ ] Please consider alternatives to the words "elaborated" and "suffers", I think you mean something slightly different.
- [ ] Please check your citation call outs in the text, in particular add white space where appropriate and check formatting.
- [ ] Replace "After compiled," with "After compilation, "Differential Evolution algorithm" by "The Differential Evolution algorithm", "In previous equation" with "In the previous equation", "P-measure" with "The P-measure", "Step function,..., pose" with "The Step function, ..., poses" etc.
- [ ] Feoktistov and Myers et al references are missing year.

Also, regarding the references:

- [ ] Please add DOIs where missing (see e.g. report by whedon)
- [ ] Please check all references, formatting, appropriate fields etc.
