### 2020-03-19 3cfc115
JOSS paper's typos fixed.
