+++ doc/JOSSPaper/paper.bib
@@ -4,7 +4,7 @@
   Author                   = {Griva, I and Nash, S G and Sofer, A},
   Publisher                = {SIAM},
   Year                     = {2009},
-  Address                  = {USA},
+  Address                  = {Philadelphia},
   Edition                  = {2}
 }
 
@@ -12,7 +12,7 @@
   Title                    = {Differential evolution: in search of solutions},
   Author                   = {Feoktistov, Vitaliy},
   Publisher                = {Springer Science+Business Media},
-  Year                     = {c2006},
+  Year                     = {2006},
   Address                  = {New York}
 }
 
@@ -45,6 +45,7 @@
   Title                    = {Head First Design Patterns},
   Author                   = {Freeman, Eric and Freeman, Elisabeth and Bates, Bert and Sierra, Kathy},
   Publisher                = {O'Reilly Media},
+  Address                  = {Sebastopol, CA},
   Year                     = {2004}
 }
 
@@ -57,8 +58,7 @@
 
 @Article{Storn:1997,
   Title                    = {Differential Evolution -- A Simple and Efficient Heuristic for global Optimization over Continuous Spaces},
-  Author                   = {Storn, Rainer
-and Price, Kenneth},
+  Author                   = {Storn, Rainer and Price, Kenneth},
   Journal                  = {Journal of Global Optimization},
   Year                     = {1997},
   Month                    = {Dec},
@@ -74,7 +74,7 @@ and Price, Kenneth},
   Author                   = {Price, K V and Storn, R M and Lampinen, J A},
   Publisher                = {Springer},
   Year                     = {2005},
-  Address                  = {Germany}
+  Address                  = {Berlin; New York}
 }
 
 @Book{Khuri:1996,
@@ -92,7 +92,7 @@ and Price, Kenneth},
   Title                    = {Response surface methodology: process and product optimization using designed experiments},
   Author                   = {Myers, Raymond H and Montgomery, Douglas C and Anderson-Cook, Christine M},
   Publisher                = {Wiley},
-  Year                     = {c2009},
+  Year                     = {2009},
   Address                  = {Hoboken, N.J.},
   Edition                  = {3}
 }
@@ -107,8 +107,10 @@ and Price, Kenneth},
 }
 
 @TechReport{Tennessee:2009,
+  Author                   = {{{M}essage {P}assing {I}nterface {F}orum}},
   Title                    = {{MPI}: {A} {M}essage-{P}assing {I}nterface {S}tandard},
   Institution              = {University of Tennessee},
+  Address                  = {Tennessee},
   Year                     = {2009}
 }
 
@@ -138,7 +140,7 @@ and Price, Kenneth},
   pages         = {{945-958}},
   month         = {{OCT}},
   issn          = {{1089-778X}},
-  doi           = {{10.1109/TEVC.2009.2014613}}
+  doi           = {10.1109/TEVC.2009.2014613}
 }
 
 @Comment{jabref-meta: databaseType:bibtex;}
