### 2020-03-19 b7bbbf7
JOSS paper bibliography's typos fixed.
