+++ doc/JOSSPaper/paper.bib
@@ -134,11 +134,10 @@
   author        = {Zhang, Jingqiao and Sanderson, Arthur C.},
   title         = {{JADE: Adaptive Differential Evolution With Optional External Archive}},
   journal       = {{IEEE Transactions on Evolutionary Computation}},
-  year          = {{2009}},
-  volume        = {{13}},
-  number        = {{5}},
-  pages         = {{945-958}},
-  month         = {{OCT}},
+  year          = {2009},
+  volume        = {13},
+  number        = {5},
+  pages         = {945-958},
   issn          = {{1089-778X}},
   doi           = {10.1109/TEVC.2009.2014613}
 }
