### 2020-03-20 8b2ba4b
JOSS paper's typos fixed.
