+++ doc/JOSSPaper/paper.md
@@ -55,7 +55,7 @@ DEPP treats the fitness function as a black-box, that must be provided by an ext
 Communication between DEPP and the external program is performed by disk files.  DEPP calls the EP by command-line, passing to it the name of a file as an argument. This file contains the name of another file (where the external program must save fitness), the dimension of the trial solution vector **X** and the components X1, X2, ... of **X**, e.g.,
 ```
 <path/filename> = arqfit: name of the file to save fitness
-        <value> =     nu: number of unknowns (dimension of the trial vector X)
+        <value> =     nu: number of unknowns (dim. of the trial vector X)
   <value of X1>
         .
         .
@@ -85,7 +85,7 @@ The following sections describe the optimization theory behind the current imple
 
 ### Differential Evolution {#sec:DE}
 
-The Differential Evolution algorithm is based on the principles of biological evolution. Basically, it works as follows. Given a population (set of discrete points in the domain of search), new individuals (trial points) are generated by “mutation” and “crossing over”. Here, “mutation” means a rule for creating new trial individuals while “crossing over” means a rule of exchanging information between a trial individual and an individual of the population. The fitness (objective function) of the new individuals is compared to the fitness of the2parent individuals and the worse individuals are eliminated. The procedure is repeated for several generations (iterations) until a criterion of convergence is satisfied within a given tolerance.
+The Differential Evolution algorithm is based on the principles of biological evolution. Basically, it works as follows. Given a population (set of discrete points in the domain of search), new individuals (trial points) are generated by “mutation” and “crossing over”. Here, “mutation” means a rule for creating new trial individuals while “crossing over” means a rule of exchanging information between a trial individual and an individual of the population. The fitness (objective function) of the new individuals is compared to the fitness of the parent individuals and the worse individuals are eliminated. The procedure is repeated for several generations (iterations) until a criterion of convergence is satisfied within a given tolerance.
 
 The current version of DEPP focus on a particular kind of maximization problem, often found in practical applications, i.e.,
 $$\begin{aligned}
@@ -104,7 +104,7 @@ After the population initialization, the fitness of each individual $\boldsymbol
 
 For the next generations, trial individuals are calculated based on the population of the previous generations. Trial individuals are created by a process of mutation and crossing over as follows. For each individual $\boldsymbol{x}_i$ of the previous population, other three distinct individuals ($\boldsymbol{x}_{i_1}$, $\boldsymbol{x}_{i_2}$ and $\boldsymbol{x}_{i_3}$) of the same population are randomly selected and a mutant individual $\boldsymbol{v}$ is created 
 $$\boldsymbol{v}=\boldsymbol{x}_{i_1}+F(\boldsymbol{x}_{i_3}-\boldsymbol{x}_{i_2}),$$ 
-where $F$ is the differentiation parameter. Then, the mutant individual crosses over with $\boldsymbol{x}_i$, that produces the trial individual $\boldsymbol{u}$. The components $u_j$ ($1 \le j \le D$) of $\boldsymbol{u}$ are given by 
+where $F$ is the differentiation parameter. Then, the mutant individual and $\boldsymbol{x}_i$ cross over, producing the trial individual $\boldsymbol{u}$. The components $u_j$ ($1 \le j \le D$) of $\boldsymbol{u}$ are given by 
 $$u_{j}=\begin{cases} 
 v_{j} &  \text{if}~~\mathcal{R} < C_r~~\text{or}~~j = j^*, \\
[…]
