### 2020-03-20 #1701 @gbertoldo
Dear, @meg-simula,

We fixed all the issues you raised, except for adding a missing DOI reported by Whedon. There is a paper with the same title as the book we cited. Whedon is claiming for the paper's DOI. Since the book has no DOI, it was not added.

We thank you and the reviewers (@HaoZeke and @dham) for the suggestions that helped us to improve the quality of the software and the quality of the manuscript.
