### 2020-03-20 #1701 @arfon
@gbertoldo - I'll take over this submission from here to help out `@meg-simula`.

One quick stylistic thing, you seem to have an inconsistency in how you refer to the software. Sometimes it's `DEPP`, other times DEPP. You might want to make sure that when you have made these all the same?

After this, could you make a new release of this software that includes the changes that have resulted from this review. Then, please make an archive of the software in Zenodo/figshare/other service and update this thread with the DOI of the archive? For the Zenodo/figshare archive, please make sure that:
