### 2020-03-20 91d3028
JOSS paper's typos fixed.
