### 2020-03-20 #1701 @gbertoldo
Hi, @arfon,

We fixed the issue you pointed out. Here is the DOI: 10.5281/zenodo.3722436.

Thanks!
