### 2020-03-20 #1701 @arfon
@gbertoldo is this DOI recommendation correct?

```
- https://doi.org/10.2307/1270613 may be missing for title: Response surface methodology: process and product optimization using designed experiments
```
